8_ Chatbot with RAG as a tool/requirements.txt
streamlit==1.55.0
# LangChain + LangGraph (compatible versions)
langchain==1.2.13
langchain-community==0.4.1
langchain-core==1.2.20  
langchain-huggingface==1.2.1

langgraph==1.1.3
langgraph-checkpoint-sqlite==3.0.3

# Embeddings + RAG
sentence-transformers==5.3.0
faiss-cpu==1.13.2

# PDF + text processing
pypdf==6.9.2

# Utilities
ddgs==9.11.4
aiosqlite==0.22.1
requests==2.32.5
python-dotenv==1.2.2
huggingface_hub==1.7.2